Added search by ID
Search by ID works only for IDs > 1000 (this is the current way of
distinguishing an (integer valued) ID from a position in the list of movies
found.

In the interactive mode, instead of entering a new search string, you can enter
something like 137523 to search the movie "Fight Club" by ID.

This is a workaround in case search by string doesn't work, and you still want
to properly tag a movie.