Using external library for movie title parsing
The old, regex-based way of parsing the movie title from the file name
was too simplistic, and it didn't work anymore with the TMDB API.
Specifically, including the year in the search string was not supported
and resulting in the offline mode becoming unusable.

I found the parse-torrent-name library, which does an admirable job for
this. Currently, the TMDB search is based solely on the title. In a
future version, I might use the year as well (if the TMDB API allows
specifying the year explicitly).