Reverted yet again to inline library
The problem was that while there was a stable 0.9 version available in
PyPi (https://pypi.org/project/tmdb/0.9/), that version has an error;
furthermore the newest version (1.0b), while listed on PyPI, is not
actually available for download anywhere anymore. Thus this commit
re-adds the dependency explicitly to the repo.

Care will need to be taken in the future to specify all the required
transitive dependencies in setup.py.