Removed setup.py dependency on module loading
The way that setup.py was reading metadata from the module was
well-intentioned, as it avoided duplicated data. The problem was,
however, that without the necessary dependencies (which are installed
by setup.py) importing the module would abort with an error.

This commit fixes the issue by hardcoding the author/e-mail metadata in
setup.py, rather than importing them from the package.